+++ libs/output-mapping/tests/Writer/Table/Strategy/LocalTableStrategyTest.php
@@ -5,22 +5,15 @@ declare(strict_types=1);
 namespace Keboola\OutputMapping\Tests\Writer\Table\Strategy;
 
 use Generator;
-use InvalidArgumentException;
 use Keboola\InputMapping\Staging\ProviderInterface;
-use Keboola\OutputMapping\Writer\Table\MappingResolver\LocalMappingResolver;
-use Keboola\OutputMapping\Writer\Table\Source\LocalFileSource;
-use Keboola\OutputMapping\Writer\Table\Source\SourceInterface;
+use Keboola\OutputMapping\Mapping\MappingFromProcessedConfiguration;
+use Keboola\OutputMapping\Tests\AbstractTestCase;
 use Keboola\OutputMapping\Writer\Table\Strategy\LocalTableStrategy;
-use Keboola\StorageApiBranch\ClientWrapper;
-use Keboola\StorageApiBranch\Factory\ClientOptions;
-use Keboola\Temp\Temp;
-use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
-use ReflectionProperty;
 use SplFileInfo;
 use Symfony\Component\Filesystem\Filesystem;
 
-class LocalTableStrategyTest extends TestCase
+class LocalTableStrategyTest extends AbstractTestCase
 {
     public function prepareLoadTaskOptionsProvider(): Generator
     {
@@ -46,14 +39,12 @@ class LocalTableStrategyTest extends TestCase
      */
     public function testPrepareLoadTaskOptions(array $config, array $expectedTags): void
     {
-        $temp = new Temp();
-        $file = $temp->createFile('test.csv');
+        $file = $this->temp->createFile('test.csv');
 
-        $clientWrapper = $this->createStorageClientWrapper();
-        $localFileMock = $this->createLocalFileSourceMock($file, false);
+        $sourceMock = $this->createSourceMock($file, false, $expectedTags);
 
         $strategy = new LocalTableStrategy(
-            $clientWrapper,
+            $this->clientWrapper,
             new NullLogger(),
             $this->createMock(ProviderInterface::class),
             $this->createMock(ProviderInterface::class),
@@ -61,7 +52,7 @@ class LocalTableStrategyTest extends TestCase
             false,
         );
 
-        $loadTaskOptions = $strategy->prepareLoadTaskOptions($localFileMock, $config);
+        $loadTaskOptions = $strategy->prepareLoadTaskOptions($sourceMock);
         self::assertCount(3, array_keys($loadTaskOptions));
 
         self::assertArrayHasKey('delimiter', $loadTaskOptions);
@@ -71,7 +62,7 @@ class LocalTableStrategyTest extends TestCase
 
         self::assertArrayHasKey('dataFileId', $loadTaskOptions);
 
-        $file = $clientWrapper->getTableAndFileStorageClient()->getFile($loadTaskOptions['dataFileId']);
+        $file = $this->clientWrapper->getTableAndFileStorageClient()->getFile($loadTaskOptions['dataFileId']);
         self::assertArrayHasKey('isSliced', $file);
         self::assertFalse($file['isSliced']);
         self::assertArrayHasKey('name', $file);
@@ -85,16 +76,13 @@ class LocalTableStrategyTest extends TestCase
      */
     public function testPrepareLoadTaskOptionsForSlicedFile(array $config, array $expectedTags): void
     {
-        $temp = new Temp();
-        $file = new SplFileInfo(sprintf('%s/myTable', $temp->getTmpFolder()));
-
+        $file = new SplFileInfo(sprintf('%s/myTable', $this->temp->getTmpFolder()));
         (new Filesystem())->mkdir($file->getPathname());
 
-        $clientWrapper = $this->createStorageClientWrapper();
-        $localFileMock = $this->createLocalFileSourceMock($file, true);
+        $sourceMock = $this->createSourceMock($file, true, $expectedTags);
 
         $strategy = new LocalTableStrategy(
-            $clientWrapper,
+            $this->clientWrapper,
             new NullLogger(),
             $this->createMock(ProviderInterface::class),
             $this->createMock(ProviderInterface::class),
@@ -102,7 +90,7 @@ class LocalTableStrategyTest extends TestCase
             false,
         );
 
-        $loadTaskOptions = $strategy->prepareLoadTaskOptions($localFileMock, $config);
+        $loadTaskOptions = $strategy->prepareLoadTaskOptions($sourceMock);
         self::assertCount(3, array_keys($loadTaskOptions));
 
         self::assertArrayHasKey('delimiter', $loadTaskOptions);
@@ -112,7 +100,7 @@ class LocalTableStrategyTest extends TestCase
 
         self::assertArrayHasKey('dataFileId', $loadTaskOptions);
 
-        $file = $clientWrapper->getTableAndFileStorageClient()->getFile($loadTaskOptions['dataFileId']);
+        $file = $this->clientWrapper->getTableAndFileStorageClient()->getFile($loadTaskOptions['dataFileId']);
         self::assertArrayHasKey('isSliced', $file);
         self::assertTrue($file['isSliced']);
         self::assertArrayHasKey('name', $file);
@@ -121,74 +109,16 @@ class LocalTableStrategyTest extends TestCase
         self::assertSame($expectedTags, $file['tags']);
     }
 
-    public function testPrepareLoadTaskOptionsFaisOnNonWorkspaceItemSource(): void
+    private function createSourceMock(SplFileInfo $file, bool $isSliced, array $tags): MappingFromProcessedConfiguration
     {
-        $strategy = new LocalTableStrategy(
-            $this->createMock(ClientWrapper::class),
-            new NullLogger(),
-            $this->createMock(ProviderInterface::class),
-            $this->createMock(ProviderInterface::class),
-            'json',
-            false,
-        );
+        $source = $this->createMock(MappingFromProcessedConfiguration::class);
 
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('Argument $source is expected to be instance of '
-            . 'Keboola\OutputMapping\Writer\Table\Source\LocalFileSource');
-
-        $strategy->prepareLoadTaskOptions($this->createMock(SourceInterface::class), []);
-    }
-
-    public function testGetters(): void
-    {
-        $dataStorageProvider = $this->createMock(ProviderInterface::class);
-
-        $metadataStorageProvider = $this->createMock(ProviderInterface::class);
-        $metadataStorageProvider->expects(self::once())
-            ->method('getPath')
-            ->willReturn('test')
-        ;
-
-        $strategy = new LocalTableStrategy(
-            $this->createMock(ClientWrapper::class),
-            new NullLogger(),
-            $dataStorageProvider,
-            $metadataStorageProvider,
-            'json',
-            false,
-        );
-
-        self::assertSame($dataStorageProvider, $strategy->getDataStorage());
-        self::assertSame($metadataStorageProvider, $strategy->getMetadataStorage());
-
-        $mappingResolver = $strategy->getMappingResolver();
-        self::assertInstanceOf(LocalMappingResolver::class, $mappingResolver);
-
-        $reflection = new ReflectionProperty($mappingResolver, 'path');
-        self::assertSame('test', $reflection->getValue($mappingResolver));
-    }
-
-    private function createStorageClientWrapper(): ClientWrapper
-    {
-        return new ClientWrapper(
-            new ClientOptions(
-                (string) getenv('STORAGE_API_URL'),
-                (string) getenv('STORAGE_API_TOKEN'),
-            ),
-        );
-    }
-
-    private function createLocalFileSourceMock(SplFileInfo $file, bool $isSliced): LocalFileSource
-    {
-        $localFileMock = $this->createMock(LocalFileSource::class);
-        $localFileMock->expects(self::once())
-            ->method('isSliced')
-            ->willReturn($isSliced)
-        ;
-        $localFileMock->expects(self::once())
-            ->method('getFile')
-            ->willReturn($file)
-        ;
-        return $localFileMock;
+        $source->expects(self::once())->method('isSliced')->willReturn($isSliced);
+        $source->expects(self::once())->method('getPathName')->willReturn($file->getPathname());
+        $source->method('getSourceName')->willReturn($file->getFilename());
+        $source->expects(self::once())->method('getTags')->willReturn($tags);
+        $source->expects(self::once())->method('getDelimiter')->willReturn(';');
+        $source->expects(self::once())->method('getEnclosure')->willReturn('|');
+        return $source;
     }
 }
