--- libs/output-mapping/tests/Writer/Table/Strategy/SqlWorkspaceTableStrategyNewTest.php
@@ -1,21 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\OutputMapping\Tests\Writer\Table\Strategy;
-
-use Keboola\InputMapping\Staging\AbstractStrategyFactory;
-use Keboola\OutputMapping\Writer\Table\Strategy\SqlWorkspaceTableStrategyNew;
-
-class SqlWorkspaceTableStrategyNewTest extends AbstractWorkspaceTableStrategyNewTestCase
-{
-    public function setUp(): void
-    {
-        parent::setUp();
-        $this->strategy = $this
-            ->getWorkspaceStagingFactory($this->clientWrapper)
-            ->getTableOutputStrategyNew(AbstractStrategyFactory::WORKSPACE_SNOWFLAKE);
-
-        self::assertInstanceOf(SqlWorkspaceTableStrategyNew::class, $this->strategy);
-    }
-}
